Avoid 'conversion may change value' cases in Splitter and SplitterGUI
Avoid a 'conversion may change value' warning in splitter.cpp by means of
separating an assignment.

Avoid three 'conversion may change value' warnings in splitter.cpp, confirming
conversions to `int` because:
- a non-Krusader function that is used (`setValue(int value)`) only accepts
that data type and a percentage is not going to be bigger than 100;
- a non-Krusader function that is used twice (`QByteArray(const char *data,
int size)`) only accepts `int` arguments.

Avoid several 'conversion to `double` may change value' warnings in
splittergui.cpp by means of:
- confirming conversions to integer data types;
- using `long double` to store the result of operations, and later converting
to `double` because that is what `QDoubleSpinBox::setValue()` accepts;
- confirming direct conversions to `double`, because no more precission is
needed, and that is what `QDoubleSpinBox::setValue()` accepts;

Avoid one 'conversion from `double` to `KIO::filesize_t`' warning in
splittergui.cpp by means of confirming conversions to integer data types.

Revision: https://invent.kde.org/utilities/krusader/-/merge_requests/57

+++ krusader/Splitter/combiner.cpp
@@ -329,7 +329,7 @@ void Combiner::combineAbortJobs()
 
 void Combiner::combineWritePercent(KJob *, unsigned long)
 {
-    auto percent = (int)((((double)receivedSize / expectedSize) * 100.) + 0.5);
+    auto percent = static_cast<int>(((static_cast<long double>(receivedSize) / expectedSize) * 100.) + 0.5);
     setValue(percent);
 }
